rename signal.h to ossignal.h to avoid shadowing C signal.h
